pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "olist-ai-data-platform"
version = "0.1.0"
description = "AI-ready e-commerce data platform MVP with multi-agent orchestration"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Data Platform Team" }
]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.1",
  "psycopg[binary]>=3.2.0",
  "psycopg-pool>=3.2.4",
  "langgraph>=0.2.34",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-google-genai>=2.0.0",
  "langsmith>=0.1.140",
  "qdrant-client>=1.12.0",
  "fastembed>=0.4.2",
  "google-genai>=0.3.0",
  "dbt-postgres>=1.8.0",
  "sqlglot>=25.17.0",
  "sse-starlette>=2.1.3",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
  "pre-commit>=4.0.0",
]

[project.scripts]
dp-api = "app.main:run_dev"
dp-ingest = "app.ingestion.loader:main"
dp-serving-refresh = "app.transform.serving:main"
dp-agent = "app.agent.core:cli_run"
dp-provision-readonly = "app.db.provision_readonly:main"
dp-rag-index = "app.rag.indexer:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
